[refactor] Together API 프롬프트를 담당하는 클래스의 @Service를 @Component로 변경 및 인터페이스 삭제
